Move Journal controller
